e2e: add tests to exercise topology manager
bootstrap the process adding initial tests.
We will add more tests in the coming days.
This is the same code which is being submitted as PR
to openshift origin, currently under review and discussion.

The tests are still being actively developed
hence we track a specific (stable) branch and we update
it before each run.

Signed-off-by: Francesco Romani <fromani@redhat.com>