tests: topology: use reservedSystemCPUs
recent 4.3 and 4.4 releases gained support for the kubelet flag
`reservedSystemCPUs` which allows the cluster admin to select
which CPU should be reserved for system tasks (e.g. running kubelet
itself!) to save cpu cores for exclusive allocation.

We make use to this option to avoid reserving CPUs on NUMA node 0,
which is most often the one on which PCI devices are attached to.

This allows us to test for full NUMA node allocation avoiding false
negatives.

Signed-off-by: Francesco Romani <fromani@redhat.com>