topology manager: temporary pull from fromani fork
We discovered a TM e2e test is broken. To merge the fix,
we need to try out in a suitable 4.5 environment, so we pull
it here to validate it - we can't make things worse for jobs
already broken.

When the fix will be merged we can revert and pull from openshift master
again.

Signed-off-by: Francesco Romani <fromani@redhat.com>